fix/make sure socket is alive before emitting

+++ client/src/app/pages/game-page/game-page.component.ts
@@ -23,7 +23,9 @@ export class GamePageComponent implements OnDestroy {
 
     ngOnDestroy() {
         const messageType = this.isHost ? 'host abandonment' : 'player abandonment';
-        this.socketService.send(messageType, this.gameService.roomId);
-        this.gameService.destroy();
+        if(this.socketService.isSocketAlive()) {
+            this.socketService.send(messageType, this.gameService.roomId);
+            this.gameService.destroy();
+        }
     }
 }
